Fixed some colour issues and text box sizes
Colours behind ScrollViewers was set to white, which I'd personally
never noticed because it is so close to our background colour -- but
that is set to our proper background colour now.

And the Text Boxes on the Volunteer Screen are now set to the correct
size. Also removed reference to the font from the first Volunteer tab -
as it's version of sizing is different from default. (This could be
reverted, and we could have this font as the new text box default?)

Removed Resources.xaml stuff from Corporate page.